+++ CHANGELOG.md
@@ -133,6 +133,13 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+## 3.0.12
+
+### Resolved Issues
+
+* \#2527 Ensure removal of docs in a `has_many` does a multi update when setting
+  to an empty array.
+
 ## 3.0.11
 
 ### Resolved Issues